feat(bm): Expose billable metric code
