+++ lago_python_client/models/plan.py
@@ -1,19 +1,8 @@
 from pydantic import BaseModel, Field
 from typing import Optional, List, Union
 
-class GroupProperties(BaseModel):
-    group_id: Optional[str]
-    values: Optional[dict]
+from .charge import Charges, ChargesResponse
 
-class Charge(BaseModel):
-    id: Optional[str]
-    billable_metric_id: Optional[str]
-    charge_model: Optional[str]
-    properties: Optional[dict]
-    group_properties: Optional[GroupProperties]
-
-class Charges(BaseModel):
-    __root__: List[Charge]
 
 class Plan(BaseModel):
     name: Optional[str]
@@ -40,6 +29,6 @@ class PlanResponse(BaseModel):
     trial_period: Optional[float]
     pay_in_advance: Optional[bool]
     bill_charges_monthly: Optional[bool]
-    charges: Optional[Charges]
+    charges: Optional[ChargesResponse]
     active_subscriptions_count: int
     draft_invoices_count: int
