+++ tests/fixtures/plan.json
@@ -17,6 +17,7 @@
       {
         "lago_id": "51c1e851-5be6-4343-a0ee-39a81d8b4ee1",
         "lago_billable_metric_id": "a6947936-628f-4945-8857-db6858ee7941",
+        "billable_metric_code": "bm_code",
         "created_at": "2022-07-01T14:47:14Z",
         "amount_currency": "EUR",
         "charge_model": "standard",
